use doubles for power comparisons

+++ src/main/java/factionsplusplus/utils/Comparators.java
@@ -5,8 +5,8 @@ import java.util.Comparator;
 import factionsplusplus.models.Faction;
 
 public class Comparators {
-    public static final Comparator<Pair<Faction, Integer>> FACTIONS_BY_POWER = (pairOne, pairTwo) -> {
-        int comparison = Integer.compare(pairOne.right(), pairTwo.right()); 
-        return Integer.compare(0, comparison);
+    public static final Comparator<Pair<Faction, Double>> FACTIONS_BY_POWER = (pairOne, pairTwo) -> {
+        int comparison = Double.compare(pairOne.right(), pairTwo.right()); 
+        return Double.compare(0, comparison);
     };
 }
